+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/Validation/Validated/ValidatedCommand.cs
@@ -9,13 +9,15 @@ namespace CleanArchitecture.TestApplication.Application.Validation.Validated
 {
     public class ValidatedCommand : IRequest, ICommand
     {
-        public ValidatedCommand(string field, string email)
+        public ValidatedCommand(string field, string email, string cascaseTest)
         {
             Field = field;
             Email = email;
+            CascaseTest = cascaseTest;
         }
 
         public string Field { get; set; }
         public string Email { get; set; }
+        public string CascaseTest { get; set; }
     }
 }