+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.BlazorClient/HttpClients/Contracts/Services/Validation/ValidatedCommandValidator.cs
@@ -23,10 +23,25 @@ namespace CleanArchitecture.TestApplication.BlazorClient.HttpClients.Contracts.S
             RuleFor(v => v.Email)
                 .NotNull()
                 .EmailAddress();
+
+            RuleFor(v => v.CascaseTest)
+                .NotNull()
+                .Cascade(CascadeMode.Stop)
+                .MinimumLength(1)
+                .CustomAsync(ValidateCascaseTestAsync);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         private async Task ValidateFieldAsync(
+            string value,
+            ValidationContext<ValidatedCommand> validationContext,
+            CancellationToken cancellationToken)
+        {
+            throw new NotImplementedException("Your custom validation rules here...");
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
+        private async Task ValidateCascaseTestAsync(
             string value,
             ValidationContext<ValidatedCommand> validationContext,
             CancellationToken cancellationToken)