+++ Modules/Intent.Modules.Application.MediatR.FluentValidation/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 4.4.4
 
 - Improvement: Removed `// IntentMatch(...)` code management instructions from templates which are no longer needed since version `4.4.0` of the `Intent.OutputManager.RoslynWeaver` module.
+- Improvement: Added CascadeMode option on Fluent Validations to specify this behaviour.
 
 ### Version 4.4.3